requirements.txt
flask
pyserial
opencv-python
numpy
